+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.4.9] unreleased
+### Added
+- You can now copy paste using box place, cut and copy also works, doesnt work for volumes or objects "yet"
+
 ### Fixed
 - While box placing it will no longer cancel if mouse is moved off canvas
 - You can now use right click to cancel a box place
+- Right click and middle mouse can no longer draw to canvas, fixed bad code
 
 ## [1.4.8]
 ### Fixed